# contracts/openapi/ttndd-ops-api.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/organizations": {
      "post": {
        "operationId": "OrgConfigController_createOrganization",
        "summary": "Create a new organization",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": ["Organization Config"],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/organizations/{slug}": {
      "get": {
        "operationId": "OrgConfigController_findBySlug",
        "summary": "Get organization by slug",
        "parameters": [
          {
            "name": "slug",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": ["Organization Config"],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/organizations/{id}/info": {
      "patch": {
        "operationId": "OrgConfigController_updateInfo",
        "summary": "Update organization info (name, logo)",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": ["Organization Config"],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/organizations/{id}/settings": {
      "patch": {
        "operationId": "OrgConfigController_updateSettings",
        "summary": "Update organization settings (JSONB)",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": ["Organization Config"],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/organizations/{id}/modules/{moduleName}": {
      "patch": {
        "operationId": "OrgConfigController_toggleModule",
        "summary": "Toggle a module on/off",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "moduleName",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": ["Organization Config"],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/organizations/{id}/branches": {
      "get": {
        "operationId": "OrgConfigController_getBranches",
        "summary": "List branches",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object"
                  }
                }
              }
            }
          }
        },
        "tags": ["Organization Config"],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "post": {
        "operationId": "OrgConfigController_createBranch",
        "summary": "Create a new branch",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": ["Organization Config"],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/organizations/{id}/branches/{branchId}": {
      "patch": {
        "operationId": "OrgConfigController_updateBranch",
        "summary": "Update a branch",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "branchId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": ["Organization Config"],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "delete": {
        "operationId": "OrgConfigController_deleteBranch",
        "summary": "Delete a branch (must have no active members)",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "branchId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": ["Organization Config"],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/organizations/{id}/units": {
      "get": {
        "operationId": "OrgConfigController_getUnits",
        "summary": "List units (optionally by branch)",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "branchId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object"
                  }
                }
              }
            }
          }
        },
        "tags": ["Organization Config"],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "post": {
        "operationId": "OrgConfigController_createUnit",
        "summary": "Create a new unit",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": ["Organization Config"],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/organizations/{id}/units/{unitId}": {
      "patch": {
        "operationId": "OrgConfigController_updateUnit",
        "summary": "Update a unit",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "unitId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": ["Organization Config"],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "delete": {
        "operationId": "OrgConfigController_deleteUnit",
        "summary": "Delete a unit (must have no active members or children)",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "unitId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": ["Organization Config"],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/organizations/{id}/members": {
      "get": {
        "operationId": "OrgConfigController_getMembers",
        "summary": "List members (paginated)",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": ["Organization Config"],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/organizations/{id}/audit-log": {
      "get": {
        "operationId": "OrgConfigController_getAuditLog",
        "summary": "View audit log",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "action",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "resource",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "from",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "to",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": ["Organization Config"],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/organizations/{orgId}/members": {
      "post": {
        "operationId": "IamController_inviteMember",
        "summary": "Invite a new member to the organization",
        "parameters": [
          {
            "name": "orgId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": ["IAM — Members & Roles"],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "TTNDD_OPS API",
    "description": "Thanh Thiếu Niên Đại Đạo — Platform REST API\n\nMulti-tenant ERP platform for DTNDD scouting operations.\nAll endpoints require Bearer token and X-Org-Id header for RLS isolation.",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
...
  }
}